pipeline/requirements.txt
pandas==1.3.2
python-dotenv==0.19.2
requests==2.26.0
click==7.1.2